Collapse small modules handling falcon data to single file
