--- fig/gateway/errors.py
@@ -1,10 +1,0 @@
-class TranslatorError(Exception):
-    pass
-
-
-class EventDataError(TranslatorError):
-    pass
-
-
-class FalconAPIDataError(TranslatorError):
-    pass
